RevenJ framework testing fixes
