+++ frameworks/CSharp/revenj/benchmark_config.json
@@ -1,30 +1,6 @@
 {
   "framework": "revenj",
   "tests": [{
-    "windows": {
-      "setup_file": "setup",
-      "json_url": "/bench/json",
-      "db_url": "/bench/db",
-      "query_url": "/bench/queries/",
-      "plaintext_url": "/bench/plaintext",
-      "update_url": "/bench/updates/",
-      "fortune_url": "/bench/fortunes",
-      "port": 8080,
-      "approach": "Realistic",
-      "classification": "Fullstack",
-      "database": "Postgres",
-      "database_os": "Linux",
-      "framework": "Revenj",
-      "language": "C#",
-      "flavor": "Microsoft",
-      "orm": "Full",
-      "platform": "None",
-      "webserver": "None",
-      "os": "Windows",
-      "display_name": "Revenj",
-      "notes": "",
-      "versus": "Revenj"
-    },
     "default": {
       "setup_file": "setup",
       "json_url": "/bench/json",
@@ -48,6 +24,30 @@
       "display_name": "Revenj",
       "notes": "",
       "versus": "Revenj"
+    },
+    "windows": {
+      "setup_file": "setup_windows",
+      "json_url": "/bench/json",
+      "db_url": "/bench/db",
+      "query_url": "/bench/queries/",
+      "plaintext_url": "/bench/plaintext",
+      "update_url": "/bench/updates/",
+      "fortune_url": "/bench/fortunes",
+      "port": 8080,
+      "approach": "Realistic",
+      "classification": "Fullstack",
+      "database": "Postgres",
+      "database_os": "Linux",
+      "framework": "Revenj",
+      "language": "C#",
+      "flavor": "Microsoft",
+      "orm": "Full",
+      "platform": "None",
+      "webserver": "None",
+      "os": "Windows",
+      "display_name": "Revenj",
+      "notes": "",
+      "versus": "Revenj"
     }
-  }]
+ }]
 }
