+++ frameworks/Java/revenj-jvm/benchmark_config.json
@@ -1,5 +1,5 @@
 {
-  "framework": "revenj",
+  "framework": "revenj-jvm",
   "tests": [{
     "default": {
       "setup_file": "setup",
