+++ frameworks/Java/revenj-jvm/deploy.sh
