+++ frameworks/Java/revenj-jvm/src/main/webapp/WEB-INF/jsp/fortunes.jsp
